+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## 3.2.0
+* Added `useIsolatedController` parameter to `LiteState` which allows to use the same controller type for many widgets but the controller instances must be different. In this case pass an instance of the controller and set `useIsolatedController` to true.
+See the example in `MultiControllerSetup` class in the example project
 ## 3.1.2
 * Add possibility to reset and dispose all controllers in one call
 just call `LiteState.resetAllControllers(dispose: true)`